Add logic to display correct windows for passwords
I have added the logic to display the correct windows depending if the
user needs to set up a password or enter it if one already exists on
record.

I have also created a new window subclass called SetupWindow which
handles the responsibility of helping the user set up a new password.

Lastly, I have also removed the CLI class because it was unnecessary.
Classes will now incorporate window objects as data members as needed.