Remove cli package
This commit removes the "cli" package from the codebase due to its
unnecessary implementation. The previous implementation utilized multiple
window classes to manage inputs, but these functions can be accomplished
more efficiently using the command line, eliminating the need for the
curses library.

In practice, the curses library is only necessary for displaying and
facilitating interaction with the inbox, specifically for visualizing and
selecting emails. By removing the "cli" package, we streamline the
processes of password entry, credential importing, and authorization
code input.

In essence, this commit simplifies the CLI by eliminating unnecessary
curses implementations. It shifts the focus of curses to enhancing the
email inbox, which will be integrated at a later stage.