Rename choices to emails and add type hinting for clarity
I have renamed the choices parameter to emails to better reflect what the
choices are. While in context of the method name, it might be clear what
the choices represent, I have opted for certainty of what they represent
by renaming it emails and adding typing hinting to relay the parameter
should be a list. The class should be more readable as a result.