Break down handle_ctrl_r method for readability
I broke down the handle_ctrl_r class into multiple methods for
readability.

- get_temp_file now handles creating a temporary file to write our email
to.

- get_email_contents now handles the responsibility of extracting the
different parts of the email and returning them in a list.

- prepare_and_send_email handles the responsibility of formatting the
email contents into a string and sending it to the desired email address.

- format_email_as_string ensures the email contents are adequately
formatted
into a string.

handle_ctrl_r is now responsible for bringing these methods together to
ensure an email is sent correctly.