Create EmailListManager class
This class was created to to ensure users enter the correct email.
Before loging into the client, users use the 'add' command to add
the email to a .txt file that contains a list of emails they'd like
to use for the client. When users attempt to login using an email,
it will be validated against emails in the file to ensure it is
correct.

This is necessary to prevent misspelling, entering the wrong email,
etc. This list also allows the client to simultaneously login to
multiple emails at the same time without having to enter multiple
emails directly into the CLI every time they wish to use the
client.